Refactor: 로고수정
