+++ src/components/Login/LoginView.tsx
@@ -1,5 +1,5 @@
 import { styled } from "styled-components";
-import { ReactComponent as Logo } from "../../assests/logo.svg";
+import { ReactComponent as Logo } from "../../assests/newLogo.svg";
 import { ReactComponent as CloseBtn } from "../../assests/CloseBtn.svg";
 import { useNavigate } from "react-router-dom";
 import SubmitForms from "./SubmitForms";
